requirements.txt
gradio
huggingface_hub
PyMuPDF
sentence-transformers
numpy
faiss-cpu